backend/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
dbus-python==1.3.2
Django==5.0.7
django-cleanup==8.1.0
djangorestframework==3.15.2
djangorestframework-gis==1.0
GDAL==3.6.2
geopandas==1.0.1
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
psycopg2==2.9.9
pyogrio==0.9.0
pyproj==3.6.1
python-dateutil==2.9.0.post0
pytz==2024.1
shapely==2.0.5
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
python-dotenv==1.0.1
